chore: move strategy metrics utils
